test: added test for xhtml line breaks
